Started implementing geometric decomposition algorithm.

+++ common/input.py
@@ -2,5 +2,5 @@ import numpy as np
 import csv
 
 def parse_txt(path) -> np.ndarray:
-    data = np.loadtxt(path)
+    data = np.loadtxt(path, dtype=int)
     return data
